requirements.txt
discord.py
pytube
beautifulsoup4
lxml
requests_html
python-dotenv